libraries/python/neural_hive_specialists/requirements.txt
# Core
pydantic>=2.5.2
structlog>=23.2.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0

# MLflow
mlflow>=2.9.0

# MongoDB
pymongo>=4.6.0
motor>=3.3.0

# Redis
redis[hiredis]>=5.0.1

# Neo4j
neo4j>=5.15.0

# Observabilidade
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-grpc>=0.42b0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
prometheus-client>=0.19.0

# Explicabilidade (opcional)
shap>=0.44.0
lime>=0.2.0.1

# Utilities
tenacity>=8.2.3
circuitbreaker>=1.4.0
python-json-logger>=2.0.7
networkx>=3.0

# Feature Extraction
sentence-transformers>=2.2.0

# Drift Monitoring
evidently>=0.4.0

# ML/Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Business Metrics & Anomaly Detection
joblib>=1.3.0  # Para persistência de modelos Isolation Forest
clickhouse-driver>=0.2.6  # Para histórico de métricas (opcional)

# Statistical Analysis for A/B Testing
scipy>=1.11.0  # Para chi-square test, t-test, Mann-Whitney U
statsmodels>=0.14.0  # Para análise estatística avançada (opcional)

# JWT Authentication
PyJWT>=2.8.0
cryptography>=41.0.0  # Para suporte a algoritmos RS256 (futuro)

# Compliance & Data Governance
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
# APScheduler>=3.10.0  # Opcional: para retention policies automáticas

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
testcontainers>=3.7.1
faker>=22.0.0
grpcio-testing>=1.60.0

# Disaster Recovery & Cloud Storage
boto3>=1.34.0  # AWS S3 para backups
google-cloud-storage>=2.14.0  # Google Cloud Storage para backups

# Neural Hive dependencies
# neuralHiveObservability @ file:///app/libraries/python/neural_hive_observability
